🐛 Fixed collision bug with pressure plates
The pressure plates would accidentally set the bool for checking collisions off.

This would happen if multiple objects where present on the pressure plate, then when one would be removed, the bool would be incorrectly set.

We just count the amount of collisions now and only remove it when there is 0.